+++ src/main/java/org/gdg/zipte_gdg/api/controller/admin/apartment/AdminReviewController.java
@@ -1,0 +1,17 @@
+package org.gdg.zipte_gdg.api.controller.admin.apartment;
+
+import lombok.RequiredArgsConstructor;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+@RequiredArgsConstructor
+@RequestMapping("/admin/review")
+public class AdminReviewController {
+
+    // 리뷰 강제 삭제 기능
+
+    // 리뷰 숨기기
+
+    // 댓글 강제 삭제하기
+}
