+++ src/main/java/org/gdg/zipte_gdg/api/controller/admin/shopping/request/ProductManagerRequest.java
@@ -1,4 +1,4 @@
-package org.gdg.zipte_gdg.api.controller.shopping.productManger.request;
+package org.gdg.zipte_gdg.api.controller.admin.shopping.request;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
