+++ src/main/java/org/gdg/zipte_gdg/api/controller/shopping/delivery/DeliveryController.java
@@ -14,11 +14,13 @@ public class DeliveryController {
 
     private final DeliveryService deliveryService;
 
+    // 배송 조회
     @GetMapping("/{deliveryId}")
     public ApiResponse<DeliveryResponse> getOne (@PathVariable("deliveryId") Long deliveryId) {
         return ApiResponse.created(deliveryService.findById(deliveryId));
     }
 
+    // 배송 관련 정보 수정
     @PutMapping("/{deliveryId}")
     public ApiResponse<DeliveryResponse> updateState(@PathVariable("deliveryId")Long deliveryId, @RequestBody DeliveryRequest deliveryRequest) {
         deliveryRequest.setId(deliveryId);
