+++ src/main/java/org/gdg/zipte_gdg/api/controller/user/MemberController.java
@@ -1,8 +1,8 @@
-package org.gdg.zipte_gdg.api.controller.user.member;
+package org.gdg.zipte_gdg.api.controller.user;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.log4j.Log4j2;
-import org.gdg.zipte_gdg.api.controller.user.member.request.MemberRequest;
+import org.gdg.zipte_gdg.api.controller.user.request.MemberRequest;
 import org.gdg.zipte_gdg.api.response.ApiResponse;
 import org.gdg.zipte_gdg.api.service.user.member.MemberService;
 import org.gdg.zipte_gdg.api.service.user.member.response.MemberResponse;
@@ -18,6 +18,7 @@ public class MemberController {
 
     private final MemberService memberService;
 
+    // 개인정보 수정하기
     @PutMapping
     public ApiResponse<MemberResponse> update(@RequestBody MemberRequest memberRequest, @AuthenticationPrincipal PrincipalDetails principalDetails){
 
@@ -25,10 +26,12 @@ public class MemberController {
         return ApiResponse.created(memberService.addAddress(memberRequest));
     }
 
+    // 마이페이지 조회하기
     @GetMapping("/myPage")
     public ApiResponse<MemberResponse> getOne(@AuthenticationPrincipal PrincipalDetails principalDetails){
         Long memberId = principalDetails.getId();
         return ApiResponse.created(memberService.getOne(memberId));
     }
 
+    // 특정 유저 블라인드 처리 하기
 }
